more memory efficient canvas
not everything is drawn on the canvas and the canvas is only as big as the viewport
this should reduce the amount of memory needed to hold the image

This will temporarily disable the screenshot feature.
The screenshot feature will in the future have to be implemented with a separate behind-the-scenes canvas.